% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicData.R
\name{clinicData}
\alias{clinicData}
\title{Description of clinicData}
\format{Here are the names of the columns and my guess at what they really are \cr
\strong{days_to_death} - how many days the patient survived (patient is dead) \cr
\strong{days_to_last_followup} - how many days that patient has survived so far (patient is alive) \cr
\strong{pathologic_grade} - How bad is this tumor \cr
\strong{pathologic_stage} - How advanced is this tumor \cr
\strong{sex} - Sex of patient \cr
\strong{age_at_diagnosis} - Age of patient}
\source{
Keith Baggerly and Roger Penn
}
\description{
I don't really know what it is but I will use it in the package
}

