% $Id: undocumented.Rd 1115 2006-12-05 18:22:28Z warnes $

\name{undocumented}
\alias{as.factor}
\alias{allele.count.2.genotype}
\alias{as.factor.allele.genotype}
\alias{as.factor.default}
\alias{as.factor.genotype}
\alias{shortsummary.genotype}
\alias{geno.as.array}
\alias{mknum}
\alias{hap}
\alias{hapshuffle}
\alias{hapenum}
\alias{hapfreq}
\alias{hapmcmc}
\alias{mourant}
\alias{hapambig}
\title{Undocumented functions}
\description{
  These functions are undocumented.  Some are internal and not intended
  for direct use.  Some are not yet ready for end users.  Others simply
  haven't been documented yet.
}
\usage{
}
\author{Gregory R. Warnes}
\keyword{ misc }
%%\keyword{genetics}

