# $Id: undocumented.Rd,v 1.4 2002/11/12 19:58:45 warnesgr Exp $
#
# $Log: undocumented.Rd,v $
# Revision 1.4  2002/11/12 19:58:45  warnesgr
# - Changes to remove warnings generated by 'R CMD check'.
#
# Revision 1.3  2002/11/08 21:07:28  warnesgr
#
# - DESCRIPTION: Updated version number and date
# - TODO: Updated todo list.
#
# Revision 1.2  2002/02/21 03:09:02  warnes
#
# - Add as.factor.genotype.
#
# Revision 1.1  2002/02/14 12:48:57  warnes
#
# - Added 'keywords' to HWE.test
# - Create undocumented.Rd to satisfy documentation requiremets for
#   objects not otherwise included in documentation
# - Add 'alais' commands noting that 'genotype' and '[.haplotype' are
#   documented here.
#
# Revision 1.1  2001/12/12 01:36:29  warneg
#
# Initial checkin.
#
# 

\name{undocumented}
\alais{ci.balance}
\alias{as.factor}
\alias{allele.count.2.genotype}
\alias{as.factor.allele.genotype}
\alias{as.factor.default}
\alias{as.factor.genotype}
\title{Undocumented functions}
\description{
  These functions are undocumented.  Some are internal and not intended
  for direct use.  Others simply haven't been documented yet.  
}
%\usage{
%}
\author{Gregory R. Warnes}
\keyword{ misc }
%\keyword{genetics}

